% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldnh.R
\name{convert_raw_to_melted_xlsx}
\alias{convert_raw_to_melted_xlsx}
\title{Convert raw text data file from the potentiostat (one file per voltammogram)
to a single spreadsheet containing all of the voltammograms in a melted
form, with four columns: potential, device, conc, current. Current is in
microamperes.}
\usage{
convert_raw_to_melted_xlsx()
}
\description{
Convert raw text data file from the potentiostat (one file per voltammogram)
to a single spreadsheet containing all of the voltammograms in a melted
form, with four columns: potential, device, conc, current. Current is in
microamperes.
}
\examples{
convert_raw_to_melted_xlsx()
}
